+++ canvas/createImageData.go
@@ -4,7 +4,7 @@ import (
 	"image/color"
 )
 
-func (el *Canvas) CreateImageData(width, height int, pixelColor color.RGBA) interface{} {
+func (el *Canvas) CreateImageData(width, height float64, pixelColor color.RGBA) interface{} {
 	imageData := el.SelfContext.Call("createImageData", width, height)
 	return imageData.Get("data")
 }
