+++ canvas/drawImage.go
@@ -47,7 +47,7 @@ import (
 //     final
 //     Golang Sintaxe: platform.drawImage(img, sx, sy, sWidth, sHeight, x, y, width,
 //                     height)
-func (el *Canvas) DrawImage(image interface{}, value ...int) {
+func (el *Canvas) DrawImage(image interface{}, value ...float64) {
 
 	if len(value) == 8 {
 		sx := value[0]
