+++ canvas/setHeight.go
@@ -1,5 +1,5 @@
 package canvas
 
-func (el *Canvas) SetHeight(height int) {
+func (el *Canvas) SetHeight(height float64) {
 	el.SelfContext.Set("height", height)
 }
