+++ canvas/setPixex.go
@@ -1,5 +1,5 @@
 package canvas
 
-func (el *Canvas) SetPixel(x, y int, pixel interface{}) {
+func (el *Canvas) SetPixel(x, y float64, pixel interface{}) {
 	el.SelfContext.Call("putImageData", pixel, x, y)
 }
