+++ canvas/setWidth.go
@@ -1,5 +1,5 @@
 package canvas
 
-func (el *Canvas) SetWidth(width int) {
+func (el *Canvas) SetWidth(width float64) {
 	el.SelfContext.Set("width", width)
 }
